fix(linter): keep trap rule off common helpers
